undoing changes

+++ src/main/java/models/FriendDto.java
@@ -1,29 +1,27 @@
 package models;
 
-import java.math.BigInteger;
-
 // import java.math.BigInteger;
 
 public class FriendDto {
     private Long id;
-    private BigInteger userId;
-    private BigInteger friendId;
+    private Long userId;
+    private Long friendId;
     public Long getId() {
         return id;
     }
     public void setId(Long id) {
         this.id = id;
     }
-    public BigInteger getUserId() {
+    public Long getUserId() {
         return userId;
     }
-    public void setUserId(BigInteger userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
-    public BigInteger getFriendId() {
+    public Long getFriendId() {
         return friendId;
     }
-    public void setFriendId(BigInteger friendId) {
+    public void setFriendId(Long friendId) {
         this.friendId = friendId;
     }
